+++ HWDescription/RD53A.h
@@ -65,7 +65,7 @@ class RD53A : public RD53
           uint8_t            pHybridId,
           uint8_t            pRD53Id,
           uint8_t            pRD53Lane,
-          uint8_t            pRD53eFuseCode,
+          uint32_t           pRD53eFuseCode,
           const std::string& fileName,
           const std::string& cfgComment);
     RD53A(const RD53A&) = delete;
